:art: cleanup some code
