+++ package.json
@@ -41,12 +41,10 @@
     "README.md"
   ],
   "dependencies": {
-    "chalk": "^2.1.0",
-    "winston": "^2.2.0"
+    "chalk": "^2.1.0"
   },
   "devDependencies": {
     "@types/chalk": "^0.4.31",
-    "@types/winston": "^2.3.5",
     "gulp": "^3.9.1",
     "gulptraum": "^2.2.2",
     "gulptraum-typescript": "^1.3.3",
